+++ docs/docs/.vuepress/config.js
@@ -1,0 +1,15 @@
+module.exports = {
+    title: 'Instant Analytics Documentation',
+    description: 'Documentation for the Instant Analytics plugin',
+    base: '/docs/instant-analytics/',
+    dest: '../../../../sites/nystudio107/web/docs/instant-analytics',
+    themeConfig: {
+        sidebar: [
+            ['/', 'Instant Analytics plugin for Craft CMS 3.x'],
+            ['/Overview', 'Instant Analytics Overview'],
+            ['/UseCases', 'Use Cases'],
+            ['/Configuring', 'Configuring Instant Analytics'],
+            ['/Using', 'Using Instant Analytics'],
+        ]
+    }
+};