+++ postcss.config.js
@@ -1,12 +1,16 @@
 module.exports = {
     plugins: [
-    	require('postcss-import'),
-    	require('postcss-extend'),
-    	require('postcss-simple-vars'),
-    	require('postcss-nested-ancestors'),
-    	require('postcss-nested'),
-    	require('postcss-hexrgba'),
-    	require('autoprefixer'),
-        require('tailwindcss')('./tailwind.config.js')
+        require('postcss-import')({
+            plugins: [
+                require('stylelint')
+            ]
+        }),
+        require('tailwindcss')('./tailwind.config.js'),
+        require('postcss-preset-env')({
+            autoprefixer: { grid: true },
+            features: {
+                'nesting-rules': true
+            }
+        })
     ]
-}
+};